修改PWM的启动顺序
if (HAL_TIM_PWM_Start(&htim3, TIM_CHANNEL_3) != HAL_OK)
这行代码在 RT-Thread 线程中（优先级 9）执行时，HAL_TIM_PWM_Start 内部会操作 TIM3 的寄存器（尤其是使能 CC3 输出），但此时 DMA 配置和 TIM 状态还未完全同步，或者写缓冲区（write buffer）导致异步总线错误（IMPRECISERR），最终升级为 Hard Fault。

+++ verify_ws2812b-demo/applications/bsp_ws2812b.c
@@ -96,7 +96,7 @@ rt_err_t ws2812b_update(void)
 
     memset(ws2812_buffer, 0, sizeof(ws2812_buffer));
 
-    if (HAL_TIM_PWM_Start_DMA(&htim3, TIM_CHANNEL_3, (uint16_t *)ws2812_buffer, DMA_BUFF_LEN) != HAL_OK)
+    if (HAL_TIM_PWM_Start_DMA(&htim3, TIM_CHANNEL_3, (uint32_t *)ws2812_buffer, DMA_BUFF_LEN) != HAL_OK)
     {
         LOG_E("DMA启动失败");
         is_updating = 0;
